[Jest] Use Moon for scoped jest execution (#262160)
Extracts the Moon-based Jest execution from scripts/check into a shared
module so that scripts/jest can also benefits from Moon's task caching
and parallel execution.

Just moving things around.

---------

(cherry picked from commit 77827fa8549b48dae4613a436d80f3543cca75db)

# Conflicts:
#	src/platform/packages/shared/kbn-test/index.ts
