+++ src/platform/packages/shared/kbn-test/index.ts
@@ -74,6 +74,18 @@ export {
 } from './src/jest/run_contract';
 export type { JestConfigResult, JestValidationResult } from './src/jest/run_contract';
 
+export {
+  runJestViaMoon,
+  parseMoonJestOutput,
+  findJestConfig,
+  JEST_CONFIG_NAMES,
+} from './src/jest/run_jest_via_moon';
+export type {
+  MoonJestResult,
+  MoonJestTaskResult,
+  JestFailedTest,
+} from './src/jest/run_jest_via_moon';
+
 export * from './src/kbn_archiver_cli';
 
 export * from '@kbn/kbn-client';
