+++ src/platform/packages/shared/kbn-test/tsconfig.json
@@ -34,6 +34,7 @@
     "@kbn/react-mute-legacy-root-warning",
     "@kbn/babel-preset",
     "@kbn/dot-text",
-    "@kbn/apm-utils"
+    "@kbn/apm-utils",
+    "@kbn/moon"
   ]
 }
